DRAGON/server/HEDGE_LORDS_SERVER/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "services"
version = "0.1.0"
description = "Services for the hedge lords price consumer"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi==0.115.8",
    "uvicorn[standard]==0.34.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4", 
    "bcrypt==4.0.1",
    "python-multipart==0.0.9",
    "pydantic[email]==2.10.6",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",
    "websockets==15.0",
    "numpy==2.2.2",
    "pandas==2.2.2",
    "email-validator==2.1.1",
    "SQLAlchemy==2.0.38",
    "asyncpg==0.30.0",
    "pydantic-settings==2.8.0",
    "delta-rest-client==1.0.12",
    "psycopg2-binary==2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.1.0",
    "isort>=5.10.1",
    "mypy>=0.931",
]

[tool.setuptools]
packages = ["services"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ["py311"]